+++ simwave/kernel/backend/c_code/forward/constant_density/3d/wave.c
@@ -114,7 +114,7 @@ double forward(f_type *u, f_type *velocity, f_type *damp,
                     f_type sum_z = coeff[0] * u[current_snapshot];
 
                     // radius of the stencil
-                    for(int ir = 1; ir <= stencil_radius; ir++){
+                    for(size_t ir = 1; ir <= stencil_radius; ir++){
                         //neighbors in the Y direction
                         sum_y += coeff[ir] * (u[current_snapshot + ir] + u[current_snapshot - ir]);
 
@@ -255,7 +255,7 @@ double forward(f_type *u, f_type *velocity, f_type *damp,
 
                 // null neumann on the left
                 if(y_before == 2){
-                    for(int ir = 1; ir <= stencil_radius; ir++){
+                    for(size_t ir = 1; ir <= stencil_radius; ir++){
                         size_t domain_offset = (i * nx + j) * ny + stencil_radius;
                         size_t next_snapshot = next_t * domain_size + domain_offset;
                         u[next_snapshot - ir] = u[next_snapshot + ir];
@@ -271,7 +271,7 @@ double forward(f_type *u, f_type *velocity, f_type *damp,
 
                 // null neumann on the right
                 if(y_after == 2){
-                    for(int ir = 1; ir <= stencil_radius; ir++){
+                    for(size_t ir = 1; ir <= stencil_radius; ir++){
                         size_t domain_offset = (i * nx + j) * ny + (ny - stencil_radius - 1);
                         size_t next_snapshot = next_t * domain_size + domain_offset;
                         u[next_snapshot + ir] = u[next_snapshot - ir];
@@ -301,7 +301,7 @@ double forward(f_type *u, f_type *velocity, f_type *damp,
 
                 // null neumann on the front
                 if(x_before == 2){
-                    for(int ir = 1; ir <= stencil_radius; ir++){
+                    for(size_t ir = 1; ir <= stencil_radius; ir++){
                         size_t domain_offset = (i * nx + stencil_radius) * ny + k;
                         size_t next_snapshot = next_t * domain_size + domain_offset;
                         u[next_snapshot - (ir * ny)] = u[next_snapshot + (ir * ny)];
@@ -317,7 +317,7 @@ double forward(f_type *u, f_type *velocity, f_type *damp,
 
                 // null neumann on the back
                 if(x_after == 2){
-                    for(int ir = 1; ir <= stencil_radius; ir++){
+                    for(size_t ir = 1; ir <= stencil_radius; ir++){
                         size_t domain_offset = (i * nx + (nx - stencil_radius - 1)) * ny + k;
                         size_t next_snapshot = next_t * domain_size + domain_offset;
                         u[next_snapshot + (ir * ny)] = u[next_snapshot - (ir * ny)];
@@ -347,7 +347,7 @@ double forward(f_type *u, f_type *velocity, f_type *damp,
 
                 // null neumann on the top
                 if(z_before == 2){
-                    for(int ir = 1; ir <= stencil_radius; ir++){
+                    for(size_t ir = 1; ir <= stencil_radius; ir++){
                         size_t domain_offset = (stencil_radius * nx + j) * ny + k;
                         size_t next_snapshot = next_t * domain_size + domain_offset;
                         u[next_snapshot - (ir * nx * ny)] = u[next_snapshot + (ir * nx * ny)];
@@ -363,7 +363,7 @@ double forward(f_type *u, f_type *velocity, f_type *damp,
 
                 // null neumann on the bottom
                 if(z_after == 2){
-                    for(int ir = 1; ir <= stencil_radius; ir++){
+                    for(size_t ir = 1; ir <= stencil_radius; ir++){
                         size_t domain_offset = ((nz - stencil_radius - 1) * nx + j) * ny + k;
                         size_t next_snapshot = next_t * domain_size + domain_offset;
                         u[next_snapshot + (ir * nx * ny)] = u[next_snapshot - (ir * nx * ny)];
