+++ simwave/kernel/backend/c_code/forward/variable_density/2d/wave.c
@@ -122,7 +122,7 @@ double forward(f_type *u, f_type *velocity, f_type *density, f_type *damp,
                 f_type fd_density_z = 0.0;
 
                 // radius of the stencil
-                for(int ir = 1; ir <= stencil_radius; ir++){
+                for(size_t ir = 1; ir <= stencil_radius; ir++){
                     //neighbors in the horizontal direction
                     sd_pressure_x += coeff_order2[ir] * (u[current_snapshot + ir] + u[current_snapshot - ir]);
                     fd_pressure_x += coeff_order1[ir] * (u[current_snapshot + ir] - u[current_snapshot - ir]);
@@ -252,7 +252,7 @@ double forward(f_type *u, f_type *velocity, f_type *density, f_type *damp,
 
             // null neumann on the left
             if(x_before == 2){
-                for(int ir = 1; ir <= stencil_radius; ir++){
+                for(size_t ir = 1; ir <= stencil_radius; ir++){
                     size_t domain_offset = i * nx + stencil_radius;
                     size_t next_snapshot = next_t * domain_size + domain_offset;
                     u[next_snapshot - ir] = u[next_snapshot + ir];
@@ -268,7 +268,7 @@ double forward(f_type *u, f_type *velocity, f_type *density, f_type *damp,
 
             // null neumann on the right
             if(x_after == 2){
-                for(int ir = 1; ir <= stencil_radius; ir++){
+                for(size_t ir = 1; ir <= stencil_radius; ir++){
                     size_t domain_offset = i * nx + (nx - stencil_radius - 1);
                     size_t next_snapshot = next_t * domain_size + domain_offset;
                     u[next_snapshot + ir] = u[next_snapshot - ir];
@@ -296,7 +296,7 @@ double forward(f_type *u, f_type *velocity, f_type *density, f_type *damp,
 
             // null neumann on the top
             if(z_before == 2){
-                for(int ir = 1; ir <= stencil_radius; ir++){
+                for(size_t ir = 1; ir <= stencil_radius; ir++){
                     size_t domain_offset = stencil_radius * nx + j;
                     size_t next_snapshot = next_t * domain_size + domain_offset;
                     u[next_snapshot - (ir * nx)] = u[next_snapshot + (ir * nx)];
@@ -312,7 +312,7 @@ double forward(f_type *u, f_type *velocity, f_type *density, f_type *damp,
 
             // null neumann on the bottom
             if(z_after == 2){
-                for(int ir = 1; ir <= stencil_radius; ir++){
+                for(size_t ir = 1; ir <= stencil_radius; ir++){
                     size_t domain_offset = (nz - stencil_radius - 1) * nx + j;
                     size_t next_snapshot = next_t * domain_size + domain_offset;
                     u[next_snapshot + (ir * nx)] = u[next_snapshot - (ir * nx)];
