+++ simwave/kernel/backend/c_code/forward/variable_density/3d/wave.c
@@ -127,7 +127,7 @@ double forward(f_type *u, f_type *velocity, f_type *density, f_type *damp,
                     f_type fd_density_z = 0.0;
 
                     // radius of the stencil
-                    for(int ir = 1; ir <= stencil_radius; ir++){
+                    for(size_t ir = 1; ir <= stencil_radius; ir++){
                         //neighbors in the Y direction
                         sd_pressure_y += coeff_order2[ir] * (u[current_snapshot + ir] + u[current_snapshot - ir]);
                         fd_pressure_y += coeff_order1[ir] * (u[current_snapshot + ir] - u[current_snapshot - ir]);
@@ -278,7 +278,7 @@ double forward(f_type *u, f_type *velocity, f_type *density, f_type *damp,
 
                 // null neumann on the left
                 if(y_before == 2){
-                    for(int ir = 1; ir <= stencil_radius; ir++){
+                    for(size_t ir = 1; ir <= stencil_radius; ir++){
                         size_t domain_offset = (i * nx + j) * ny + stencil_radius;
                         size_t next_snapshot = next_t * domain_size + domain_offset;
                         u[next_snapshot - ir] = u[next_snapshot + ir];
@@ -294,7 +294,7 @@ double forward(f_type *u, f_type *velocity, f_type *density, f_type *damp,
 
                 // null neumann on the right
                 if(y_after == 2){
-                    for(int ir = 1; ir <= stencil_radius; ir++){
+                    for(size_t ir = 1; ir <= stencil_radius; ir++){
                         size_t domain_offset = (i * nx + j) * ny + (ny - stencil_radius - 1);
                         size_t next_snapshot = next_t * domain_size + domain_offset;
                         u[next_snapshot + ir] = u[next_snapshot - ir];
@@ -324,7 +324,7 @@ double forward(f_type *u, f_type *velocity, f_type *density, f_type *damp,
 
                 // null neumann on the front
                 if(x_before == 2){
-                    for(int ir = 1; ir <= stencil_radius; ir++){
+                    for(size_t ir = 1; ir <= stencil_radius; ir++){
                         size_t domain_offset = (i * nx + stencil_radius) * ny + k;
                         size_t next_snapshot = next_t * domain_size + domain_offset;
                         u[next_snapshot - (ir * ny)] = u[next_snapshot + (ir * ny)];
@@ -340,7 +340,7 @@ double forward(f_type *u, f_type *velocity, f_type *density, f_type *damp,
 
                 // null neumann on the back
                 if(x_after == 2){
-                    for(int ir = 1; ir <= stencil_radius; ir++){
+                    for(size_t ir = 1; ir <= stencil_radius; ir++){
                         size_t domain_offset = (i * nx + (nx - stencil_radius - 1)) * ny + k;
                         size_t next_snapshot = next_t * domain_size + domain_offset;
                         u[next_snapshot + (ir * ny)] = u[next_snapshot - (ir * ny)];
@@ -370,7 +370,7 @@ double forward(f_type *u, f_type *velocity, f_type *density, f_type *damp,
 
                 // null neumann on the top
                 if(z_before == 2){
-                    for(int ir = 1; ir <= stencil_radius; ir++){
+                    for(size_t ir = 1; ir <= stencil_radius; ir++){
                         size_t domain_offset = (stencil_radius * nx + j) * ny + k;
                         size_t next_snapshot = next_t * domain_size + domain_offset;
                         u[next_snapshot - (ir * nx * ny)] = u[next_snapshot + (ir * nx * ny)];
@@ -386,7 +386,7 @@ double forward(f_type *u, f_type *velocity, f_type *density, f_type *damp,
 
                 // null neumann on the bottom
                 if(z_after == 2){
-                    for(int ir = 1; ir <= stencil_radius; ir++){
+                    for(size_t ir = 1; ir <= stencil_radius; ir++){
                         size_t domain_offset = ((nz - stencil_radius - 1) * nx + j) * ny + k;
                         size_t next_snapshot = next_t * domain_size + domain_offset;
                         u[next_snapshot + (ir * nx * ny)] = u[next_snapshot - (ir * nx * ny)];
